# [RFC] metanode: Load snapshot file with multiple goroutines