--- a/cli/cmd/const.go
+++ b/cli/cmd/const.go
@@ -73,6 +73,7 @@ const (
 	CliFlagDelWorkerSleepMs   = "delete-worker-sleep-ms"
 	CliFlagMarkDelRate        = "mark-delete-rate"
 	CliFlagCrossZone          = "crossZone"
+	CliFlagMetaFileBlock      = "mf-blksize"
 
 	//CliFlagSetDataPartitionCount	= "count" use dp-count instead