--- a/master/const.go
+++ b/master/const.go
@@ -34,6 +34,7 @@ const (
 	dataPartitionSizeKey    = "size"
 	metaPartitionCountKey   = "mpCount"
 	volCapacityKey          = "capacity"
+	metaFileBlockKey        = "metaFileBlock"
 	volOwnerKey             = "owner"
 	volAuthKey              = "authKey"
 	replicaNumKey           = "replicaNum"