--- a/master/gapi_volume.go
+++ b/master/gapi_volume.go
@@ -173,9 +173,9 @@ func (s *VolumeService) volPermission(ctx context.Context, args struct {
 }
 
 func (s *VolumeService) createVolume(ctx context.Context, args struct {
-	Name, Owner, ZoneName, Description                     string
-	Capacity, DataPartitionSize, MpCount, DpReplicaNum     uint64
-	FollowerRead, Authenticate, CrossZone, DefaultPriority bool
+	Name, Owner, ZoneName, Description                                string
+	Capacity, DataPartitionSize, MpCount, DpReplicaNum, MetaFileBlock uint64
+	FollowerRead, Authenticate, CrossZone, DefaultPriority            bool
 }) (*Vol, error) {
 	uid, per, err := permissions(ctx, ADMIN|USER)
 	if err != nil {
@@ -191,7 +191,7 @@ func (s *VolumeService) createVolume(ctx context.Context, args struct {
 	}
 
 	vol, err := s.cluster.createVol(args.Name, args.Owner, args.ZoneName, args.Description, int(args.MpCount),
-		int(args.DpReplicaNum), int(args.DataPartitionSize), int(args.Capacity),
+		int(args.DpReplicaNum), int(args.DataPartitionSize), int(args.Capacity), int(args.MetaFileBlock),
 		args.FollowerRead, args.Authenticate, args.CrossZone, args.DefaultPriority)
 	if err != nil {
 		return nil, err